refactor(gatekeeper): move GatekeeperCompletion model to llm module
