+++ src/linux_mcp_server/gatekeeper/llm.py
@@ -2,20 +2,30 @@
 
 import logging
 
+from pydantic import BaseModel
+
 from linux_mcp_server.config import CONFIG
 from linux_mcp_server.config import GatekeeperProvider
-from linux_mcp_server.gatekeeper.anthropic_client import complete_anthropic
-from linux_mcp_server.gatekeeper.gemini_client import complete_gemini
-from linux_mcp_server.gatekeeper.openai_client import complete_openai
-from linux_mcp_server.gatekeeper.openrouter_client import complete_openrouter
-from linux_mcp_server.gatekeeper.vertex_ai_client import complete_vertex_ai
-from linux_mcp_server.models import GatekeeperCompletion
 
 
 logger = logging.getLogger("linux-mcp-server")
 
 
+class GatekeeperCompletion(BaseModel):
+    text: str
+    prompt_tokens: int = 0
+    completion_tokens: int = 0
+    usage_cost: float | None = None
+
+
 async def complete_gatekeeper(prompt: str, *, max_tokens: int) -> GatekeeperCompletion:
+    # Deferred imports avoid circular dependency: clients import GatekeeperCompletion from this module.
+    from linux_mcp_server.gatekeeper.anthropic_client import complete_anthropic
+    from linux_mcp_server.gatekeeper.gemini_client import complete_gemini
+    from linux_mcp_server.gatekeeper.openai_client import complete_openai
+    from linux_mcp_server.gatekeeper.openrouter_client import complete_openrouter
+    from linux_mcp_server.gatekeeper.vertex_ai_client import complete_vertex_ai
+
     assert CONFIG.gatekeeper is not None
     provider = CONFIG.gatekeeper.provider
     match CONFIG.gatekeeper.provider:
